orderbot/pyproject.toml
[project]
name = "orderbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentevals>=0.0.9",
    "fastapi[standard]>=0.128.0",
    "google-adk>=1.25.1",
    "httpx>=0.28.1",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "supabase>=2.11.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
